Remove Xorg
The idea itself is good, but alas it could not be realized to the end. In its current form it is just a piece of useless junk, it belongs in a landfill, not in an application. I will not continue to develop it, it would be easier to port Xorg from Termux.